requirements.txt
asgiref==3.7.2
certifi==2023.7.22
configparser==6.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-minio-storage==0.3.10
django-s3-file-field==0.3.2
djangorestframework==3.14.0
minio==6.0.2
pathvalidate==2.3.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.4